requirements_xpu.txt
# Core project dependencies
python-telegram-bot==20.8
flask==2.3.3
flask-sqlalchemy==3.1.1
langchain==0.2.0
chromadb==0.5.0
sentence-transformers==2.2.2
python-dotenv==1.0.1

# Intel XPU optimized dependencies
--pre --upgrade ipex-llm[xpu] --extra-index-url https://pytorch-extension.intel.com/release-whl/stable/xpu/us/
transformers==4.42.0
accelerate==0.29.3
bitsandbytes==0.43.0

# PyTorch with XPU support
--pre torch==2.7.1+xpu torchvision==0.22.1+xpu torchaudio==2.7.1+xpu --index-url https://download.pytorch.org/whl/nightly/xpu

# Utilities
rich==13.7.1
tqdm>=4.66.6